Add example using external process - Project property added for Docker/Jar/External process mode switch - Readme updated
