+++ src/test/java/com/github/nagyesta/lowkeyvault/example/AzureAccessTestDockerConfiguration.java
@@ -29,7 +29,7 @@ public class AzureAccessTestDockerConfiguration implements DisposableBean {
     private final LowkeyVaultContainer lowkeyVaultContainer;
 
     public AzureAccessTestDockerConfiguration() {
-        final DockerImageName imageName = DockerImageName.parse("nagyesta/lowkey-vault:1.1.0");
+        final DockerImageName imageName = DockerImageName.parse("nagyesta/lowkey-vault:1.2.0");
         lowkeyVaultContainer = new LowkeyVaultContainer(imageName, Collections.emptySet())
                 .withImagePullPolicy(PullPolicy.defaultPolicy())
                 .withExposedPorts(8443);
